+++ src/impls/maps/btree_map.rs
@@ -3,7 +3,14 @@ use alloc::collections::BTreeMap;
 use fluent_result::expect::dbg::ExpectNone;
 
 use crate::errors::{CollectionError, Collision};
-use crate::{TryExtend, TryExtendSafe, TryFromIterator};
+use crate::{TryExtend, TryExtendOne, TryExtendSafe, TryFromIterator};
+
+fn try_extend_basic<K: Ord, V, I>(map: &mut BTreeMap<K, V>, iter: &mut I) -> Result<(), Collision<(K, V)>>
+where
+    I: Iterator<Item = (K, V)>,
+{
+    iter.try_for_each(|kvp| map.try_extend_one(kvp))
+}
 
 impl<K: Ord, V, I> TryFromIterator<I> for BTreeMap<K, V>
 where
@@ -16,14 +23,11 @@ where
     /// See [trait level documentation](trait@TryFromIterator) for an example.
     fn try_from_iter(into_iter: I) -> Result<Self, Self::Error> {
         let mut iter = into_iter.into_iter();
-        iter.try_fold(Self::new(), |mut map, (k, v)| match map.contains_key(&k) {
-            true => Err((map, (k, v))),
-            false => {
-                map.insert(k, v).expect_none("should not be occupied");
-                Ok(map)
-            }
-        })
-        .map_err(|(map, kvp)| CollectionError::collision(iter, map, kvp))
+        let mut map = Self::new();
+        match try_extend_basic(&mut map, &mut iter) {
+            Ok(()) => Ok(map),
+            Err(err) => Err(CollectionError::new(iter, map, err)),
+        }
     }
 }
 
@@ -38,14 +42,7 @@ where
     /// See [trait level documentation](trait@TryExtend) for an example.
     fn try_extend(&mut self, iter: I) -> Result<(), Self::Error> {
         let mut iter = iter.into_iter();
-        iter.try_for_each(|(key, value)| match self.contains_key(&key) {
-            true => Err((key, value)),
-            false => {
-                self.insert(key, value).expect_none("should not be occupied");
-                Ok(())
-            }
-        })
-        .map_err(|kvp| CollectionError::collision(iter, Self::new(), kvp))
+        try_extend_basic(self, &mut iter).map_err(|err| CollectionError::new(iter, Self::new(), err))
     }
 }
 
@@ -59,17 +56,14 @@ where
     fn try_extend_safe(&mut self, iter: I) -> Result<(), Self::Error> {
         let mut iter = iter.into_iter();
         iter.try_fold(Self::new(), |mut map, (key, value)| match self.contains_key(&key) {
-            true => Err((map, (key, value))),
-            false => match map.contains_key(&key) {
-                true => Err((map, (key, value))),
-                false => {
-                    map.insert(key, value).expect_none("should not be occupied");
-                    Ok(map)
-                }
+            true => Err((map, Collision::new((key, value)))),
+            false => match map.try_extend_one((key, value)) {
+                Ok(()) => Ok(map),
+                Err(err) => Err((map, err)),
             },
         })
         .map(|map| self.extend(map))
-        .map_err(|(map, kvp)| CollectionError::collision(iter, map, kvp))
+        .map_err(|(map, err)| CollectionError::new(iter, map, err))
     }
 }