+++ src/lib.rs
@@ -1,10 +1,7 @@
 #![doc = include_str!("../README.md")]
-#![warn(clippy::pedantic)]
-#![warn(clippy::nursery)]
-#![warn(clippy::cargo)]
-#![warn(missing_docs)]
-#![allow(clippy::match_bool)]
-#![allow(clippy::single_match_else)]
+#![warn(clippy::pedantic, clippy::cargo, clippy::nursery)]
+#![warn(missing_docs, missing_debug_implementations)]
+#![allow(clippy::match_bool, clippy::single_match_else)]
 #![cfg_attr(not(feature = "std"), no_std)]
 
 #[cfg(feature = "std")]